requirements.txt
alpaca-py==0.43.2
pandas>=2.0.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
matplotlib>=3.8.0